adjusting data structures and transformation

+++ messages/Sailboat.go
@@ -4,5 +4,5 @@ package messages
 type Sailboat struct {
 	Name      string   `json:"Name"`
 	Position  Vector2D  `json:"Position"`
-	Timestamp float32  `json:"Timestamp"`
+	Timestamp float64  `json:"Timestamp"`
 }
