+++ messages/Speedboat.go
@@ -5,5 +5,5 @@ type Speedboat struct {
 	Name      string   `json:"Name"`
 	Position  Vector2D `json:"Position"`
 	Velocity  Vector2D `json:"Velocity"`
-	Timestamp float32  `json:"Timestamp"`
+	Timestamp float64  `json:"Timestamp"`
 }
